Project/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.13.1
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
Pillow==9.2.0